+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - add CLI to generate api-def from openapi spec
 - add `reconfigure` method to `Api` & `Endpoint` to allow for changing options after creation
 - support for path params surrounded by curly braces, e.g. `/users/{id}`
+- add `updateBody` to `RequestContext` to allow for changing the body in middleware
 
 # 0.11.1